Add files via upload
I have developed an automation framework using Python, Selenium, and Behave to interact with the bikroy.com website. This framework facilitates automated testing and interaction with various features of the bikroy.com platform. By leveraging Selenium for web automation and Behave for behavior-driven development (BDD), the framework provides a structured approach to testing functionality and ensuring the smooth operation of the website. With this automation suite uploaded to GitHub, developers and testers can easily access and utilize the codebase to enhance testing processes, improve efficiency, and maintain the quality of bikroy.com's web services